+++ src/subgraph/index.ts
@@ -29,6 +29,7 @@ export const selectAuctiontype = (id: string, auctions: Auction[]): auctionType 
 }
 
 // change this to check for other units like USDC and USDT
+// refactor this function when adam finishes subgraph
 export const generateInitialAuctionData = async (
   // eslint-disable-next-line
   auctionBidsRequest: Promise<any>,
