+++ src/utils/Defaults/index.tsx
@@ -1,5 +1,6 @@
 // Externals
 import dayjs, { Dayjs } from 'dayjs'
+import { BigNumber, BigNumberish } from 'ethers'
 
 // Interface
 import { Auction, auctionType } from 'src/interfaces/Auction'
@@ -100,3 +101,47 @@ export const resizeWindow = (x: number, y: number) => {
 export const addHours = (dayjsInstance: Dayjs, hours: number) => dayjsInstance.clone().add(hours, 'h')
 export const utcDate = dayjs(new Date().toUTCString())
 export const dateUTC = dayjs.unix(utcDate.unix())
+
+// helpers
+export const getZeros = (decimals: number) => {
+  let zeros = '0'
+  while (zeros.length < 256) {
+    zeros += zeros
+  }
+  if (decimals >= 0 && decimals <= 256 && !(decimals % 1)) {
+    return '1' + zeros.substring(0, decimals)
+  }
+
+  throw new Error('invalid decimal')
+}
+
+export const convertToNumber = (number: BigNumberish, decimals = 18): number => {
+  // big number checks & convert if not
+  let value: BigNumberish = BigNumber.from(number)
+
+  const addedZeros = getZeros(decimals)
+
+  // negative check
+  const negative = value.lt(BigNumber.from(0))
+  if (negative) {
+    value = value.mul(BigNumber.from(-1))
+  }
+
+  let fraction = value.mod(addedZeros).toString()
+  while (fraction.length < addedZeros.length - 1) {
+    fraction = '0' + fraction
+  }
+
+  // eslint-disable-next-line
+  fraction = fraction.match(/^([0-9]*[1-9]|0)(0*)/)![1]
+
+  const whole = value.div(addedZeros).toString()
+
+  value = whole + '.' + fraction
+
+  if (negative) {
+    value = '-' + value
+  }
+
+  return Number(value)
+}
